refactor and move to translator
